# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API EBA gestion documents de projet",
        "description": "Cette API REST EST une **application web pour la gestion électronique des documents liés aux projets d'EBA (e-Business Afrique).**",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "CSRF": {
                "type": "apiKey",
                "in": "header",
                "name": "remember_token"
            }
        }
    },
    "security": [
        {
            "CSRF": []
        }
    ],
    "paths": {
        "/history/list": {
            "get": {
                "tags": [
                    "History requests actions"
                ],
                "summary": "List history",
                "description": "Affiche la liste complète des historiques.",
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/history/delete": {
            "delete": {
                "tags": [
                    "History requests actions"
                ],
                "summary": "Delete histories",
                "description": "Supprime tous l'historique.",
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "User requests actions"
                ],
                "summary": "Create user",
                "description": "Créer un utilisateur.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "fullname": "ibraum",
                                    "email": "ibraumodnokpro@gmail.com",
                                    "password": "ibraum#1"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User requests actions"
                ],
                "summary": "Login user",
                "description": "Connexion d'un utilisateur.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "email": "ibraumodnokpro@gmail.com",
                                    "password": "ibraum#1"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/user/list": {
            "get": {
                "tags": [
                    "User requests actions"
                ],
                "summary": "List users",
                "description": "Affiche la liste de tous les utilisateurs dans la base de données.",
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/project/create": {
            "post": {
                "tags": [
                    "Project requests actions"
                ],
                "summary": "Create project",
                "description": "Créer un projet.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "libel": "bobilion",
                                    "client_name": "ibraum",
                                    "begin_date": "2023-12-03",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/project/list": {
            "get": {
                "tags": [
                    "Project requests actions"
                ],
                "summary": "List project",
                "description": "Affiche la liste de tous les projets.",
                "security": [
                    {
                        "CSRF": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}